requirements.txt
fastapi
uvicorn[standard]
bcrypt>=4.0.0
python-jose[cryptography]
python-multipart
